restructure solution into three projects and add NLog
