--- web-api/Api/Data/Dtos/StudentDto.cs
@@ -1,9 +1,0 @@
-﻿namespace Api.Data.Dtos
-{
-    public class StudentDto
-    {
-        public string Name { get; set; } = null!;
-        public string? Mail { get; set; }
-        public bool IsActive { get; set; }
-    }
-}
