+++ web-api/StudentCompass.Services/Contracts/IAcademicService.cs
@@ -1,0 +1,6 @@
+﻿namespace StudentCompass.Services.Contracts
+{
+    public interface IAcademicService
+    {
+    }
+}
