+++ web-api/StudentCompass.Services/Services/AcademicService.cs
@@ -1,0 +1,8 @@
+﻿using StudentCompass.Services.Contracts;
+
+namespace StudentCompass.Services.Services
+{
+    public class AcademicService : IAcademicService
+    {
+    }
+}
