+++ web-api/StudentCompass.Web/Helpers/ServiceRegistration.cs
@@ -1,14 +1,20 @@
-﻿using Api.Services.Contracts;
-using Api.Services.Repositories;
+﻿using StudentCompass.Data.Contracts;
+using StudentCompass.Data.Repositories;
+using StudentCompass.Services.Contracts;
+using StudentCompass.Services.Services;
 
-namespace Api.Services.Helpers
+namespace StudentCompass.Web.Helpers
 {
     public static class ServiceRegistration
     {
         public static IServiceCollection AddRepositories(this IServiceCollection services)
         {
+            // Repositories
             services.AddScoped<IStudentRepository, StudentRepository>();
             services.AddScoped<IAcademicRepository, AcademicRepository>(); // todo: singleton ... ?
+
+            // Services
+            services.AddScoped<IAcademicService, AcademicService>();
 
             return services;
         }
